requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-filter==23.2
django-markdown==0.8.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Markdown==3.4.3
mysqlclient==2.1.1
Pillow==9.5.0
Pygments==2.15.1
PyJWT==2.7.0
pyserial==3.5
pytz==2023.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
Serializer==0.2.1
simplejson==3.19.1
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
tzdata==2023.3
urllib3==2.0.3
